Refactor Windows error resolution for frontend startup.
