+++ frontend/src/utils/auth.ts
@@ -1,5 +1,5 @@
 import logger from '../../logger.config.mjs';
-import { OnIncompletePaymentFoundType } from '@/constants/types';
+import { OnIncompletePaymentFoundType } from '@/constants/pi';
 
 export const onIncompletePaymentFound : OnIncompletePaymentFoundType = payment => {
   logger.info('onIncompletePaymentFound:', { payment });
